feat: test 코드 리셋
