+++ src/test/java/com/sprint/mission/discodeit/security/AuthControllerTest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.security.controller;
+package com.sprint.mission.discodeit.security;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
